Refactored answers out of RatingsGuideQuestion component
This simplifies the question component, and gives the RatingsGuideAssessment component more of a purpose.
